LayoutTests/http/tests/inspector/filesystem/get-metadata-expected.txt

Tests getMetadata.

errorCode: 0
metadata:
  modificationTime: (exists)
  size: 4
errorCode: 0
metadata:
  modificationTime: (exists)
  size: 0
